Renamed tool to JinjaGen
Renamed the tool to JinjaGen to diversify it from codegen
